requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
sumy==0.11.0
openai==1.35.13
python-dotenv==1.0.1
GoogleNews==1.6.14